requirements.txt
apache_beam[gcp]
tensorflow_transform
tensorflow-data-validation
pandas
seaborn
matplotlib
datalab
google-cloud-bigquery==1.6.0
google-cloud-pubsub
pandas-gbq
google_compute_engine
mlflow
bs4